Modificações feitas para adicionar Sidebar
O que foi feito:
Adicionar a sidebar em todas as features de avaliação
Redirecionamento de páginas (ao clicar em voltar, redirecionar a uma página específica)
Centralização de cards
Estilização e organização dos botões
Adicionar link na página de avaliações, ao clicar em uma avaliação específica, o usuário consegue visualizar os detalhes
