+++ package.json
@@ -22,14 +22,17 @@
   "dependencies": {
     "@hookform/resolvers": "^3.1.0",
     "axios": "^1.4.0",
+    "bootstrap": "^5.3.3",
     "react": "^18.2.0",
+    "react-bootstrap": "^2.10.9",
     "react-dom": "^18.2.0",
     "react-hook-form": "^7.44.1",
+    "react-icons": "^5.5.0",
     "react-router-dom": "^6.11.2",
     "zod": "^3.21.4"
   },
   "devDependencies": {
-    "@badeball/cypress-cucumber-preprocessor": "^18.0.1",
+    "@badeball/cypress-cucumber-preprocessor": "^22.0.1",
     "@bahmutov/cypress-esbuild-preprocessor": "^2.2.0",
     "@cucumber/cucumber": "^9.2.0",
     "@cypress/code-coverage": "^3.10.7",
@@ -45,8 +48,8 @@
     "@typescript-eslint/parser": "^5.57.1",
     "@vitejs/plugin-react": "^4.0.0",
     "@vitest/coverage-c8": "^0.31.1",
-    "cypress": "^12.15.0",
-    "esbuild": "^0.18.8",
+    "cypress": "^14.2.0",
+    "esbuild": "^0.25.1",
     "eslint": "^8.38.0",
     "eslint-plugin-react-hooks": "^4.6.0",
     "eslint-plugin-react-refresh": "^0.3.4",
