+++ src/app/home/styles/ObterReview.module.css
@@ -11,8 +11,9 @@
     justify-content: space-between;
     margin-bottom: 12px;
     padding: 8px;
-    background-color: #f9f9f9;
+    background-color: #fff;
     border-radius: 8px;
+    word-wrap: break-word;
   }
   
   .label {
@@ -20,11 +21,28 @@
     color: #444;
   }
   
+  .comentario {
+    max-height: 200px;
+    overflow-y: auto;
+    word-wrap: break-word;
+    white-space: pre-wrap;
+  }
+  
   .footer {
     display: flex;
     justify-content: center;
+    align-items: center;
+    width: 100%;
     margin-top: 25px;
+    gap: 15px; /* Espaçamento entre os botões */
   }
+  
+  .button {
+    font-size: 16px;
+    padding: 12px 72px;
+    margin: 0;
+  }
+  
   
   .error {
     color: #d32f2f;
@@ -57,4 +75,32 @@
     justify-content: center;
     margin: 10px 0;
   }
+
+  .pageContainer {
+    display: flex;
+    justify-content: center;
+    align-items: flex-start;
+    min-height: 100vh;
+    padding: 2rem;
+    background-color: #fff;
+  }
+  
+  .layoutContainer {
+    display: flex;
+    width: 100%;
+    max-width: 1200px;
+  }
+  
+  .sidebarWrapper {
+    width: 250px;
+  }
+  
+  .contentWrapper {
+    flex: 1;
+    padding: 2rem;
+    display: flex;
+    flex-direction: column;
+    align-items: center;
+  }
+
